shopify_like_django/requirements.txt
django==4.2.14
djangorestframework==3.15.2
mysql-connector-python==9.0.0
djangorestframework-simplejwt==5.2.0
